+++ src/dummy_data/InfoCardData.json
@@ -1,0 +1,23 @@
+{
+    "title": "أحدث المشاريع",
+    "data": [
+      {
+        "title": "هاكاثون المستقبل النسخة الثانية",
+        "subTitle": "حلول تقنية تساعد الملتحقين بالجامعة من طلاب وأعضاء هيئة التدريس",
+        "type": "announcement",
+        "key": 1
+      },
+      {
+        "title": "هاكاثون المستقبل النسخة الثانية",
+        "subTitle": "حلول تقنية تساعد الملتحقين بالجامعة من طلاب وأعضاء هيئة التدريس",
+        "type": "announcement",
+        "key": 2
+      },
+      {
+        "title": "هاكاثون المستقبل النسخة الثانية",
+        "subTitle": "حلول تقنية تساعد الملتحقين بالجامعة من طلاب وأعضاء هيئة التدريس",
+        "type": "announcement",
+        "key": 3
+      }
+    ]
+  }
